+++ app/assets/javascripts/home/HomeInsights.js
@@ -3,7 +3,7 @@ import React from 'react';
 import Card from '../components/Card';
 import CheckStudentsWithLowGrades from './CheckStudentsWithLowGrades';
 import CheckStudentsWithHighAbsences from './CheckStudentsWithHighAbsences';
-import {inExperienceTeam} from '../helpers/PerDistrict';
+import {shouldShowLowGradesBox} from '../helpers/PerDistrict';
 
 
 // On the home page, show users the answers to their most
@@ -13,7 +13,7 @@ class HomeInsights extends React.Component {
     const {educatorId, educatorLabels} = this.props;
     return (
       <div className="HomeInsights" style={styles.root}>
-        {inExperienceTeam(educatorLabels) &&
+        {shouldShowLowGradesBox(educatorLabels) &&
           <CheckStudentsWithLowGrades educatorId={educatorId} />}
         <CheckStudentsWithHighAbsences educatorId={educatorId} />
         {this.renderPlaceholder()}
