+++ app/models/educator_label.rb
@@ -8,7 +8,7 @@ class EducatorLabel < ActiveRecord::Base
     uniqueness: { scope: [:label_key, :educator] },
     inclusion: {
       in: [
-        'shs_experience_team',
+        'shs_experience_team', # deprecated
         'k8_counselor',
         'high_school_house_master',
         'class_list_maker_finalizer_principal',
@@ -19,7 +19,30 @@ class EducatorLabel < ActiveRecord::Base
     }
   }
 
-  def self.has_label?(educator_id, label_key)
+  # Static labels are set by records in the database.  Dynamic labels are
+  # computed at read time here based on some other property of the educator.
+  def self.labels(educator)
+    static_labels = educator.educator_labels.map(&:label_key)
+    dynamic_labels = self.dynamic_labels_for_educator(educator)
+
+    static_labels + dynamic_labels
+  end
+
+  def self.dynamic_labels_for_educator(educator)
+    dynamic_labels = []
+
+    authorizer = Authorizer.new(educator)
+    authorized_sections = educator.sections.select do |section|
+      authorizer.is_authorized_for_section?(section)
+    end
+    if authorized_sections.size > 0
+      dynamic_labels << 'should_show_low_grades_box'
+    end
+
+    dynamic_labels
+  end
+
+  def self.has_static_label?(educator_id, label_key)
     EducatorLabel.find_by(educator_id: educator_id, label_key: label_key).present?
   end
 end
